refactor: group roommodal and form
