+++ frontend/src/components/RoomModal/index.tsx
@@ -1,6 +1,6 @@
 import { X } from "lucide-react";
 import { useCallback, useRef, useState } from "react";
-import ConfirmationDialog from "./ConfirmationDialog";
+import ConfirmationDialog from "../ConfirmationDialog";
 import RoomForm from "./RoomForm";
 import "./RoomModal.css";
